index.ejs and app.js
